ETCM-166: Move crypto to separate project
