+++ bytes/src/main/scala/io/iohk/ethereum/utils/ByteStringUtils.scala
@@ -1,0 +1,12 @@
+package io.iohk.ethereum.utils
+
+import akka.util.ByteString
+
+object ByteStringUtils {
+  def hash2string(hash: ByteString): String =
+    Hex.toHexString(hash.toArray[Byte])
+
+  // unsafe
+  def string2hash(hash: String): ByteString =
+    ByteString(Hex.decode(hash))
+}
