+++ crypto/src/main/scala/io/iohk/ethereum/crypto/SymmetricCipher.scala
