+++ crypto/src/test/scala/io/iohk/ethereum/crypto/SecureRandomBuilder.scala
@@ -1,0 +1,7 @@
+package io.iohk.ethereum.crypto
+
+import java.security.SecureRandom
+
+private[crypto] trait SecureRandomBuilder {
+  lazy val secureRandom: SecureRandom = new SecureRandom()
+}
